refactor: Serivce 로또 기능단위로 분리
