+++ src/main/java/lotto/service/LottoInfoService.java
@@ -7,7 +7,12 @@ import lotto.domain.WinningLotto;
 import java.sql.Connection;
 import java.sql.SQLException;
 
-public class DBGetter {
+public class LottoInfoService {
+
+    public static Integer getCurrentRound(Connection connection) throws SQLException {
+        RoundDAO roundDAO = new RoundDAO(connection);
+        return roundDAO.getCurrentRound();
+    }
 
     public static Lottos getDBLottos(Connection connection, String lottoRound) throws SQLException {
         LottoDAO lottoDAO = new LottoDAO(connection);
@@ -22,15 +27,5 @@ public class DBGetter {
     public static LottoResultDTO getDBLottoResultDTO(Connection connection, String lottoRound) throws SQLException {
         LottoResultDAO lottoResultDAO = new LottoResultDAO(connection);
         return lottoResultDAO.findByLottoRound(lottoRound);
-    }
-
-    public static Integer getCurrentRound(Connection connection) throws SQLException {
-        RoundDAO roundDAO = new RoundDAO(connection);
-        return roundDAO.getCurrentRound();
-    }
-
-    public static Integer getNextRound(Connection connection) throws SQLException {
-        RoundDAO roundDAO = new RoundDAO(connection);
-        return roundDAO.getNextRound();
     }
 }
