+++ src/main/java/lotto/service/LottoResultService.java
@@ -1,27 +1,14 @@
 package lotto.service;
 
-import lotto.dao.LottoDAO;
 import lotto.dao.LottoResultDAO;
-import lotto.dao.RoundDAO;
 import lotto.dao.WinningLottoDAO;
 import lotto.domain.LottoResult;
-import lotto.domain.Lottos;
 import lotto.domain.WinningLotto;
 
 import java.sql.Connection;
 import java.sql.SQLException;
 
-public class DBLoader {
-
-    public static void loadDBRoundTable(Connection connection, Integer round) throws SQLException {
-        RoundDAO roundDAO = new RoundDAO(connection);
-        roundDAO.addRound(round.toString());
-    }
-
-    public static void loadDBLottoTable(Connection connection, Integer round, Lottos lottos) throws SQLException {
-        LottoDAO lottoDAO = new LottoDAO(connection);
-        lottoDAO.addLottos(round.toString(), lottos);
-    }
+public class LottoResultService {
 
     public static void loadDBWinningLottoTable(Connection connection, Integer round, WinningLotto winningLotto)
             throws SQLException {
